+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on the Steamclock [Release Management Guide](https://github.
 ## Unreleased 
 - Added ThrowableBlocker functional interface to allow specific Throwables/Exceptions to be suppressed from being sent as an error to the crash reporting destination (#22)
 - Update Sentry from 4.3.0 to 5.1.0
+- [1] Add log rotation functionality 
 
 ---